requirements.txt
deta==1.1.0
pandas==1.4.4
plotly==5.9.0
python-dotenv==0.21.0
streamlit==1.16.0
streamlit_extras==0.2.5
